Fixed bug with no dropdown at start
